+++ src/main/container/main.go
@@ -2,11 +2,16 @@ package container
 
 import (
 	create_user_use_case "go-api/src/core/useCases/create-user"
+	create_user_producer_use_case "go-api/src/core/useCases/create-user-producer"
 	get_user_use_case "go-api/src/core/useCases/get-user"
 	list_users "go-api/src/core/useCases/list-user"
+
+	create_user_amqp "go-api/src/infra/amqp/producer/user-create"
 	users_cache "go-api/src/infra/cache/users"
 	json_place_holder "go-api/src/infra/http/integrations/jsonplaceholder"
 	model_user "go-api/src/infra/repositories/user"
+
+	amqp_client "go-api/src/main/module/amqp/rabbitmq/client"
 	cache_client "go-api/src/main/module/cache/redis"
 	http_service "go-api/src/main/module/http/client"
 	"os"
@@ -20,9 +25,10 @@ type (
 	}
 
 	Container struct {
-		GetUserUseCase    get_user_use_case.GetUserUseCase
-		CreateUserUseCase create_user_use_case.CreateUserUseCase
-		ListUsersUseCase  list_users.ListUsersUseCase
+		GetUserUseCase            get_user_use_case.GetUserUseCase
+		CreateUserUseCase         create_user_use_case.CreateUserUseCase
+		CreateUserProducerUseCase create_user_producer_use_case.CreateUserUseCase
+		ListUsersUseCase          list_users.ListUsersUseCase
 	}
 )
 
@@ -31,6 +37,10 @@ func NewContainerConfig(db *gorm.DB) *ContainerConfig {
 }
 
 func NewContainer(container_config *ContainerConfig) *Container {
+	//amqp injection
+	amqp_client := amqp_client.New()
+	create_user_amqp := create_user_amqp.NewProdocer(amqp_client)
+
 	//integration injection
 	http_service := http_service.New()
 	json_place_holder_integration := json_place_holder.New(http_service, os.Getenv("JSON_PLACE_OLDER_INTEGRATION_URL"))
@@ -50,6 +60,7 @@ func NewContainer(container_config *ContainerConfig) *Container {
 		CreateUserUseCase: create_user_use_case.NewUseCase(
 			user_repository,
 		),
-		ListUsersUseCase: list_users.NewUseCase(json_place_holder_integration, users_cache),
+		ListUsersUseCase:          list_users.NewUseCase(json_place_holder_integration, users_cache),
+		CreateUserProducerUseCase: create_user_producer_use_case.NewUseCase(create_user_amqp),
 	}
 }