+++ src/main/module/amqp/rabbitmq/server/rabbitmq-server-interface.go
@@ -1,15 +1,14 @@
-package amqp_server
+package rabbitmq
 
 import (
 	"go-api/src/main/container"
 	consumer "go-api/src/presentation/amqp/consumers"
 )
 
-type AmqpServer interface {
-	New(c *container.Container) AmqpServer
+type IAmqpServer interface {
+	New(c *container.Container) IAmqpServer
 	Start()
 	StartConsumers(container []consumer.Comsumer, i int)
 	NeedToReconnect(err error, msg string)
 	LoadConsumers(c *container.Container) []consumer.Comsumer
-	LoadProducers(c *container.Container)
 }